requirements.txt
PyQt5
GitPython
requests
python-docx
reportlab
pyperclip
pyinstaller